shared, folder 페이지 파일 구조 및 기초 틀 제작
