+++ src/components/folderList/FolderData.jsx
@@ -1,0 +1,8 @@
+import React from "react";
+import "./FolderData.css";
+
+function FolderData({ name }) {
+  return <li className="folderData">{name}</li>;
+}
+
+export default FolderData;
