+++ src/components/folderList/folderListData.css
@@ -1,0 +1,6 @@
+.folderList {
+  display: flex;
+  flex-direction: row;
+  list-style: none;
+  gap: 8px;
+}
